+++ Cosette/Engine/Ai/Score/Evaluators/KingSafetyEvaluator.cs
@@ -6,6 +6,33 @@ namespace Cosette.Engine.Ai.Score.Evaluators
 {
     public static class KingSafetyEvaluator
     {
+        private static ulong[][] _kingSafetyMask;
+
+        static KingSafetyEvaluator()
+        {
+            _kingSafetyMask = new ulong[2][];
+
+            for (var color = 0; color < 2; color++)
+            {
+                var offset = color == Color.White ? 8 : -8;
+                _kingSafetyMask[color] = new ulong[64];
+
+                for (var fieldIndex = 0; fieldIndex < 64; fieldIndex++)
+                {
+                    var mask = BoxPatternGenerator.GetPattern(fieldIndex);
+                    var fieldIndexWithOffset = fieldIndex + offset;
+
+                    if (fieldIndexWithOffset >= 0 && fieldIndexWithOffset < 64)
+                    {
+                        mask |= BoxPatternGenerator.GetPattern(fieldIndexWithOffset);
+                        mask &= ~(1ul << fieldIndex);
+                    }
+
+                    _kingSafetyMask[color][fieldIndex] = mask;
+                }
+            }
+        }
+
         public static int Evaluate(BoardState board, int openingPhase, int endingPhase, ulong fieldsAttackedByWhite, ulong fieldsAttackedByBlack)
         {
             return Evaluate(board, Color.White, openingPhase, endingPhase, fieldsAttackedByBlack) - 
@@ -16,7 +43,7 @@ namespace Cosette.Engine.Ai.Score.Evaluators
         {
             var king = board.Pieces[color][Piece.King];
             var kingField = BitOperations.BitScan(king);
-            var fieldsAroundKing = BoxPatternGenerator.GetPattern(kingField);
+            var fieldsAroundKing = _kingSafetyMask[color][kingField];
 
             var attackedFieldsAroundKing = fieldsAroundKing & fieldsAttackedByEnemy;
             var attackersCount = (int)BitOperations.Count(attackedFieldsAroundKing);